import of ServerInfoBean added to restore
For some reason, the restore page stopped working. Manual import of the
bean, like this, solves the issue. It has probably something to do with
my setup and some spring magic I don't understand. But I believe this
import is correct to do anyway. It is imported in index.jsp, create.jsp
and autoonvert.jsp already.