Implemented custom global attributes
The user can press a button to add custom attributes in the general
metadata tab.

Removed harcoded checks for "required" metadata while refactoring
checkAndExposeNext to also include checking that added fields

The custom attribute names are following CF for naming convention: can
only start with a letter, and can only contain letters, numbers and
underscores. Further specified letters to be a-zA-Z only.

CSS: removed absolute position of the ul elements, and set the height 
for jWizard to auto with a minimum. This solves an issue with bad
looking ui for when the added general metadata list gets long.